pkg/proc,service: support linux/386 (#1884)
Implement debugging function for 386 on linux with reference to AMD64.
There are a few remaining problems that need to be solved in another time.

1. The stacktrace of cgo are not exactly as expected.
2. Not implement `core` for now.
3. Not implement `call` for now. Can't not find `runtime·debugCallV1` or
   similar function in $GOROOT/src/runtime/asm_386.s.

Update #20

+++ pkg/proc/linutil/regs_test.go
@@ -46,7 +46,7 @@ func TestAMD64Get(t *testing.T) {
 		t.Fatal(err)
 	}
 	if eax != 0xDEADBEEF {
-		t.Fatalf("expected %#v, got %#v\n", 0xdeadbeef, eax)
+		t.Fatalf("expected %#v, got %#v\n", uint64(0xdeadbeef), eax)
 	}
 
 	// Test RAX, full 64 bits of register
